[fix] norminette (still working)
