+++ new/src/wrap/write.c
@@ -6,7 +6,7 @@
 /*   By: enogaWa <enogawa@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/12 20:24:50 by yshimoda          #+#    #+#             */
-/*   Updated: 2023/03/15 12:58:26 by enogaWa          ###   ########.fr       */
+/*   Updated: 2023/03/16 12:01:47 by enogaWa          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
